requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.4
django-phonenumber-field==8.0.0
fonttools==4.55.2
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.2.0
packaging==24.2
phonenumbers==8.13.51
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.2
tk==0.1.0
tzdata==2024.2
